backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# LangChain and LangGraph
langgraph==0.0.40
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langchain-google-genai==0.0.5

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1

# OpenAI for LLM and Voice
openai==1.3.0

# ElevenLabs for TTS
elevenlabs==0.2.24

# Deepgram for STT (alternative)
deepgram-sdk==2.12.0

# Audio processing
pydub==0.25.1
numpy==1.24.3

# Vector DB
pinecone-client==2.2.4

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
python-jose[cryptography]==3.3.0

# Utilities
httpx==0.25.2
